Depend on JSON::PP, not JSON and cache version checks
Will use JSON if already loaded, otherwise will use JSON::PP.  This
is a good compromise that offers potential speed if JSON::XS is available
but only requires JSON::PP as a minimum, as that is what has been
approved to go into the Perl core.

As a slight efficiency, YAML::Tiny and JSON::PP version checks
are cached, instead of calling ->VERSION every time.